refactor: share lakehouse testable editor/state and restore ingest testing styles
